made check_invoices a util function, so they are also checked when getting the promo code status
