+++ lib/devcon_signup.php
@@ -1,14 +1,15 @@
 <?php
 	require_once 'bitpay/bp_lib.php';
 	require_once 'db_config.php';
+	require_once 'check_open_invoices.php';
 
 	$adminMail = ""; //the address the confirmation mail is being sent from
 	$basic_price = 279;
 
 	//connect to mysql server
 	$mysqli = new mysqli($db_host, $db_username, $db_password, $db_name);
-	check_open_invoices($mysqli);
 	//$mysqli->report_mode = MYSQLI_REPORT_ALL;
+	check_open_invoices($mysqli);
 
 	if ( isset($_POST['email']) AND isset($_POST['firstName']) AND isset($_POST['lastName']) ) {
 		// creating the db entry for the signup
@@ -152,53 +153,6 @@
 			    'X-Mailer: PHP/' . phpversion();
 
 			mail($to, $subject, $message, $headers);
-		}
-	}
-
-	function check_open_invoices($mysqli) {
-		$time = time();
-		$sql = "SELECT * FROM signups WHERE status = 'new' AND expirationtime < ".$time;
-
-		$stmt = $mysqli->prepare($sql);
-		$stmt->execute();
-		$result=$stmt->get_result();
-		$unredeemedCodes = array();
-		while ( $signup = mysqli_fetch_array($result, MYSQLI_ASSOC) ) {
-			$invoice = bpGetInvoice($signup['invoiceid']);
-			if ( array_key_exists('status', $invoice) ) {
-				$status = $invoice['status'];
-
-				// check if a promocode was involved and should be reset because it wasn't redeemed
-				if ( $signup['promocode'] != NULL && ( $status == 'expired' || $status == 'invalid' ) ) {
-					if ( !array_key_exists($signup['promocode'], $unredeemedCodes) ) {
-						$unredeemedCodes[$signup['promocode']] = 1;
-					} else {
-						$unredeemedCodes[$signup['promocode']]++;
-					}
-				}
-
-				// update the signup status
-				$id = $signup['id'];
-				$sql = "UPDATE
-			                signups
-			            SET
-			                status = '$status'
-			            WHERE
-			            	id = $id";
-			    $mysqli->query($sql);
-			}
-		}
-		$stmt->close();
-
-		// now decrease all codes according to the unused codes in unpaid invoices
-		foreach ($unredeemedCodes as $code => $invalidate) {
-		    $sql = "UPDATE
-		    			promocodes
-		    		SET
-		    			amount_invalidated = amount_invalidated + $invalidate
-		    		WHERE
-		    			code = '$code'";
-		    $mysqli->query($sql);
 		}
 	}
 ?>
